pyproject.toml
[project]
name = "toxic_language"
version = "0.0.0"
description = "Validates whether the generated text is toxic."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "nltk",
    "transformers",
    "torch",
    "detoxify"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]